# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User and Game Catalog API",
    "description": "API para registro, login, listagem de usuários e um catálogo de jogos com autenticação JWT.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Game": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "titulo": {
            "type": "string"
          },
          "genero": {
            "type": "string"
          },
          "AnoLancamento": {
            "type": "integer"
          }
        }
      }
    }
  },
  "paths": {
    "/api/user/register": {
      "post": {
        "tags": ["Users"],
        "summary": "Registrar um novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Erro ao registrar o usuário"
          }
        }
      }
    },
    "/api/user/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro ao realizar login"
          }
        }
      }
    },
    "/api/user/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Obter a lista de usuários",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro ao obter a lista de usuários"
          }
        }
      }
    },
    "/api/games": {
      "get": {
        "tags": ["Games"],
        "summary": "Listar todos os jogos",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de jogos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Game" }
                }
              }
            }
          },
          "401": {
            "description": "Token de autenticação ausente ou inválido"
          }
        }
      },
      "post": {
        "tags": ["Games"],
        "summary": "Criar um novo jogo",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Game"
              },
              "example": {
                "titulo": "The Legend of Zelda",
                "genero": "Aventura",
                "anoLancamento": 1986
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Jogo criado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Game" }
              }
            }
          },
          "400": {
            "description": "Dados inválidos para criação do jogo"
          },
          "401": {
            "description": "Token de autenticação ausente ou inválido"
          }
        }
      }
    },
    "/api/games/{id}": {
      "get": {
        "tags": ["Games"],
        "summary": "Obter um jogo por ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jogo encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Game" }
              }
            }
          },
          "404": {
            "description": "Jogo não encontrado"
          },
          "401": {
            "description": "Token de autenticação ausente ou inválido"
          }
        }
      },
      "put": {
        "tags": ["Games"],
        "summary": "Atualizar um jogo existente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Game"
              },
              "example": {
                "titulo": "The Legend of Zelda: Breath of the Wild",
                "genero": "Aventura",
                "anoLancamento": 2017
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Jogo atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Game" }
              }
            }
          },
          "400": {
            "description": "Dados inválidos para atualização do jogo"
          },
          "404": {
            "description": "Jogo não encontrado"
          },
          "401": {
            "description": "Token de autenticação ausente ou inválido"
          }
        }
      },
      "delete": {
        "tags": ["Games"],
        "summary": "Excluir um jogo",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Jogo excluído com sucesso"
          },
          "404": {
            "description": "Jogo não encontrado"
          },
          "401": {
            "description": "Token de autenticação ausente ou inválido"
          }
        }
      }
    }
  }
}